Move properties into api package
