+++ src/main/java/net/skinsrestorer/api/SkinsRestorerAPI.java
@@ -25,7 +25,7 @@ import lombok.Getter;
 import net.skinsrestorer.shared.exception.SkinRequestException;
 import net.skinsrestorer.shared.storage.SkinStorage;
 import net.skinsrestorer.shared.utils.MojangAPI;
-import net.skinsrestorer.shared.utils.property.IProperty;
+import net.skinsrestorer.api.property.IProperty;
 
 /**
  * API Example: https://github.com/SkinsRestorer/SkinsRestorerAPIExample
