+++ src/main/java/net/skinsrestorer/api/bungeecord/events/SkinApplyBungeeEvent.java
@@ -26,7 +26,7 @@ import lombok.Setter;
 import net.md_5.bungee.api.connection.ProxiedPlayer;
 import net.md_5.bungee.api.plugin.Cancellable;
 import net.md_5.bungee.api.plugin.Event;
-import net.skinsrestorer.shared.utils.property.IProperty;
+import net.skinsrestorer.api.property.IProperty;
 import org.jetbrains.annotations.NotNull;
 
 @Getter
