+++ src/main/java/net/skinsrestorer/api/property/VelocityProperty.java
@@ -19,24 +19,27 @@
  * <http://www.gnu.org/licenses/gpl-3.0.html>.
  * #L%
  */
-package net.skinsrestorer.shared.utils.property;
+package net.skinsrestorer.api.property;
 
-import com.mojang.authlib.properties.Property;
+import com.velocitypowered.api.util.GameProfile.Property;
 import lombok.ToString;
 
 @ToString
-public class BukkitProperty extends Property implements IProperty {
-    public BukkitProperty(String name, String value) {
-        super(name, value);
-    }
+public class VelocityProperty implements IProperty {
+    private final Property property;
 
-    public BukkitProperty(String name, String value, String signature) {
-        super(name, value, signature);
+    public VelocityProperty(String name, String value, String signature) {
+        property = new Property(name, value, signature);
     }
 
     @Override
     public Object getHandle() {
-        return this;
+        return property;
+    }
+
+    @Override
+    public String getName() {
+        return property.getName();
     }
 
     @Override
@@ -45,8 +48,18 @@ public class BukkitProperty extends Property implements IProperty {
     }
 
     @Override
+    public String getValue() {
+        return property.getValue();
+    }
+
+    @Override
     public void setValue(String value) {
         throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public String getSignature() {
+        return property.getSignature();
     }
 
     @Override
