+++ src/main/java/net/skinsrestorer/bukkit/SkinsRestorer.java
@@ -42,8 +42,8 @@ import net.skinsrestorer.shared.utils.*;
 import net.skinsrestorer.shared.utils.log.LoggerImpl;
 import net.skinsrestorer.shared.utils.log.SRLogger;
 import net.skinsrestorer.shared.utils.log.console.BukkitConsoleImpl;
-import net.skinsrestorer.shared.utils.property.GenericProperty;
-import net.skinsrestorer.shared.utils.property.IProperty;
+import net.skinsrestorer.api.property.GenericProperty;
+import net.skinsrestorer.api.property.IProperty;
 import org.bstats.bukkit.Metrics;
 import org.bstats.charts.SingleLineChart;
 import org.bukkit.Bukkit;
