+++ src/main/java/net/skinsrestorer/bungee/commands/SkinCommand.java
@@ -40,7 +40,7 @@ import net.skinsrestorer.shared.storage.CooldownStorage;
 import net.skinsrestorer.shared.storage.Locale;
 import net.skinsrestorer.shared.utils.C;
 import net.skinsrestorer.shared.utils.log.SRLogger;
-import net.skinsrestorer.shared.utils.property.IProperty;
+import net.skinsrestorer.api.property.IProperty;
 
 import java.util.concurrent.TimeUnit;
