+++ src/main/java/net/skinsrestorer/shared/storage/SkinStorage.java
@@ -27,8 +27,8 @@ import net.skinsrestorer.shared.exception.SkinRequestException;
 import net.skinsrestorer.shared.utils.C;
 import net.skinsrestorer.shared.utils.MojangAPI;
 import net.skinsrestorer.shared.utils.log.SRLogger;
-import net.skinsrestorer.shared.utils.property.GenericProperty;
-import net.skinsrestorer.shared.utils.property.IProperty;
+import net.skinsrestorer.api.property.GenericProperty;
+import net.skinsrestorer.api.property.IProperty;
 
 import javax.sql.RowSet;
 import java.io.*;
